implement querying of current prices and history
As a first step towards an autonomously acting bot, this change
introduces the functionality to query the current price of the EUR/USD
currency pair as well as 20 values of the past with a granularity of 4
hours. Based on these data points, we calculate the Exponential Moving
Average for all 20 values and 10 values, respectively.