introduce __queryWidths method
By now we have two functions that print data in a tabularized format
with variable column widths. To simplify creation of such a tablular
view and not require every new user to reinvent the same error-prone
mechanism, this change introduces a new function __queryWidths that
automates the process a bit more.