add EventStreamer and RateStreamer classes
This change adds two new classes to the project: EventStreamer can be
used to subscribe to a stream of events coming from the Oanda server.
Events include the triggering of stop losses, the cancelation of an
order and more. RateStreamer on the other hand subscribes to a stream of
price changes for a set of instruments/currencies. This way we can get
frequent updates on the current prices and can react in a timely manner.