register signal handler to tear down gracefully on Ctrl+C
When terminating a streaming program abruptly we often see a backtrace
indicating an unhandled interruption in some HTTP connection.
This change registers a signal handler for SIGINT which is sent (among
others) when a user presses Ctrl+C. In this case we first disconnect the
streams and then exit normally.
As a bonus, this change also introduces a new command line parameter:
-v/--verbose. This parameter is meant to increase the verbosity of
output throughout the program.