handle SIGTERM signals along with SIGINT
A previous change installed a custom signal handler to handle SIGINT
signals which are usually sent to a program when Ctrl+C is pressed. A
less common way to shutdown a process is to directly send SIGTERM to it
on which it can react.
This change also registers the same signal handler introduced previously
for handling SIGTERM signals and correctly shutting down the
infrastructure.