introduce ThreadedStreamer class
The current approach of creating and polling on event/tick streams for
new data from the main program object does not work correctly under all
circumstances. For starters, waiting on the streams is a blocking
operation which means we can only wait on one stream (a select-like
mechanism is not available). In addition, signal handling is screwed up
once the main thread is waiting in a blocking system call.
For these reasons, we need to perform the waiting for new data on the
two streams in threads running concurrently to the main thread. This
change introduces the ThreadedStreamer class that moves the
functionality of OANDA's Streamer object into a separate thread. For
transfer of data (i.e., work items) between the stream threads and the
main thread, a queue object is used.