introduce worker class to handle work asynchronously
Python's signal handling infrastructure is really a pain. To the best of
my knowledge it is impossible (without directly interfacing with
operating system APIs not directly accessible from within Python) to get
correct signal handling when blocking system calls are in use. Although
Python provides the set_wakeup_fd() functionality which can be used to
unblock a thread waiting in a poll or select call, doing so in
conjunction with higher level APIs is impossible (because one cannot
hook into the lower level system calls used).
To that end, the main thread must not execute any blocking system calls
if a program is to react to signals in a timely manner. This change
introduces a Worker class that pushes the waiting for and processing of
events previously done in the main thread into another thread. Using
this mechanism we can now finally get correct signal handling while
still only waking up when necessary.