modify the format of the logging output
This change modifies the format of log output as logged using the
logging module. Previously (using the default format), no timestamp was
contained in the output and the component/logger to which the request
was issued was printed before the message.
With this change we use a format of the form:
  "<date> <time> <log lvl> <msg>"
where date and time are in a user readable format and log level is one
of [DD], [II], [WW], [EE], [XX], representing the default log levels.