introduce --currencies parameter
This change introduces a --currencies parameter that accepts a list of
currencies. In this first change, this list of currencies can be used to
influence the output as produced by the --list-currencies option: only
currencies supplied will be listed in this case.
Later on, the list of currencies will influence other functionality as
well.