introduce a watchdog class
During testing it could repeatedly be witnessed that sometimes the
stream tracking the current prices of currencies just delivers nothing.
This case cannot easily detected since no error is reported (my
suspicion is that for some reason the server does not want to serve our
stream).
This change introduces a watchdog thread that periodically wakes up,
checks if no price updates have been received and in that case
explicitly queries the prices and creates a new tick. Note that the
checking whether or not price updates have been received is still to be
done. For now, sending a tick happens unconditionally.
Using such a watchdog also paves the way to providing an option to not
using the streams at all, which might be preferable under certain
circumstances.