add module for handling times as per RFC3339
With this change a new module is introduced: datetimeRfc3339 can be used
to parse dates encoded as specified in RFC3339. Note that currently only
the relevant subset for our purposes is supported: only times encoded
for UTC timezone (ending with a 'Z').
This change also introduces a unit test to verify the parsing is correct
based on some strings actually received and some made up ones.