rethink price handling
With this change prices -- at the moment represented by a dict -- are
always transfered in a strongly typed fashion, i.e., the keys
'instrument', 'time', 'ask', and 'bid' are of types string, datetime
object, and Decimal, respectively.
The parsing functionality for the rateStreamer was directly added into
it and so such a stronger typed object can directly be passed to the
worker.