employ a caching scheme for a currency's historic data
With this change the history() method of Currency objects does not
necessarily query the Oanda server on every invocation. Instead, we
store the last data we retrieved and decide whether it is still current
based on the granularity of the request, the current time, as well as
the last time data was actually retrieved from the server.
The overall goal towards which this change is a first step is to reduce
the overall amount of connections to the Oanda servers as much as
possible since there exists a limit of 4 (peak 5) connections per
second. If this limit is exceeded clients might be subject to throttling
or other means degrading the quality of service.