move API object creation into fxBot.py
In this change the creation of the API object representing the
interaction point with OANDA's server is moved from the Program class
into the fxBot.py file containing the main() function.
This refactoring marks a first step towards a possible virtualization of
the API interface so that we can better test our code by supplying a
mock API object.