introduce Server abstraction
This change introduces a new abstraction to the program: a Server object
is used for interacting with the OANDA servers. That is, it abstracts
away from the direct usage of the API class provided by the oandapy
module.
The two streamer classes rateStreamer and eventStreamer are unaffected
by this currently but they are already a form of abstraction over direct
usage of the corresponding oandapy Streamer class.