add unit test for Currency class
Having introduced the abstraction of a Server class, we are now able to
create custom Server objects. One opportunity opened by this feature is
creation of mock objects to fake expected behavior or virtualize
interfaces.
This change introduces a unit test for the Currency class and its
caching functionality. A MockServer object is used to simulate the
Server interface to Currency objects while not actually performing any
network traffic but just working on local data.