update README file to reflect new mock dependency
A recent change added a unit test for the Currency class. Currency
employs a cache that is time dependent. To make the tests works
reliable, the datetime.now() method was mocked.
To accomplish this mocking the mock module was used which is now a new
dependency. This change updates the README.md file to reflect this
requirement.