introduce Price class
Before this change we had the problem that prices were displayed with as
many post decimal positions as they have or as the currently set
precision allows for. This leads to confusingly large values being
printed.
This change introduces the Price class. Objects of this class have an
associated 'pip' value that quantifies the number of relevant post
decimal positions to print. Prices use Decimal internally so as to not
loose any precision.
They support the basic arithmetic operations +, -, *, and / and interact
well with Decimal objects. Prices are usually used in conjunction with
currencies, in other cases usage of Decimal objects might be favorable.