add complete infrastructure test
This change adds a unit test that verifies the basic infrastructure for
pushing events from the various sources (rate and event streamer
threads, and watchdog) works as intended.
Focus is on execising all main paths so as to catch potential type
errors and the like. Correct evaluation of the events has to be verified
by a unit test for each component.