propagate timeout value to use in watchdog from fxBot.py
Until now the timeout value to use in the watchdog was hardcoded in the
class itself. At some point we want to make it user configurable. As an
intermediate step, this change propagates the value from fxBot.py down
into watchdog.py where it is actually used. The value is still hardcoded
in fxBot.py, however.