add unit test for automated testing of Program class
This change adds a new unit test for testing the Program class. The test
currently execises all paths that compile a list of output and terminate
(i.e., all but the endless mode).
Note that the output itself is not verified since it is subject to
change and we really do not care that much. Focus is laid instead on
really execising the paths to catch type errors and the like.