have exceptions in child threads tear down program
When one of the used threads (worker, watchdog, rate streamer, or event
streamer) raises an uncaught exception the thread itself is terminated
but the program keeps running. This behavior is not desired: an uncaught
exception in one of the threads should terminate the application after
printing some information.
This change implements the desired behavior. A newly introduced
decorator ('tryRun') is used to guard all functions running in separate
threads. Once it catches an exception, a trace back is printed and
SIGTERM send to the main thread, which will then terminate the
application gracefully.