terminate application if streams encounter an error
Up until now when one of the streams (RateStreamer or EventStreamer)
encounteres an error (signalled to it via its 'on_error' method), it
will print the error and disconnect the stream. The thread continues
running as if nothing happened.
This change modifies the behavior to fail fast: in case of an error we
raise an exception which will terminate the thread and, in fact, the
whole application.
This approach may potentially be too harsh for certain error conditions.
However, until now all witnessed cases where the 'on_error' method
was invoked were all errors that really should lead to program
termination (when the supplied access token is invalid, for instance).