introduce proxy functionality and TimeProxy class
This change introduces a new abstraction for implementing functionality
related to interaction with a Server object: a proxy. A proxy is an
object that implements a subset of a Server's methods and can intercept
the requests before the server was invoked and the responses after it
finished. Proxies can be stacked in which case the stacking order
defines the order in which requests are passed to the respective proxy
instance.
Along with the basic proxy feature, the TimeProxy class is added.
TimeProxy can be used to estimate a server's time. It does so by
investigating the server's responses to certain requests which are known
to carry them time along with them. Once acquired, it relates these
times to the system's local time to calculate a delta. This delta (or
better: the average of a series of deltas gathered over time) is used to
give an estimation of the server's time based on the system's current
local time.