introduce CacheProxy class
A previous change introduced caching functionality for the Currency
class. Every history() invocation checked whether a cached response
exists that could be used to satisfy the request.
Although working, caching is better done accross all requests that
are directed to a server.
This change introduces a new CacheProxy class: a proxy that implements
our desired caching behavior. At the moment, caching happens only for
history() invocations. It is very well possible to extent it to apply to
other requests as well, although different drawbacks must be considered.