introduce LimitProxy class
According to documentation regarding the OANDA APIs a maximum of 4
connections per second to one of the servers is allowed by a single
client. More connections will result in responses being throttled. No
exact details are given as to how long throttling continues or whether
potentially other measures are taken as well.
We do not want to take chances and so this change introduces the
LimitProxy, a proxy that ensures that no more than 4 requests are sent
to the server in each second. Along with the class, a unit test is added
that verifies the intended behavior.