add unit test to verify _Proxy wraps all Server methods
We have the problem that the proxy functionality requires manual
instruction on which of the Server's methods to wrap. This may cause
problems when the Server class is extended with new methods but the
_Proxy is forgotten.
This change adds a unit test to check for this case: If the Server gets
a new method the unit test will fail until the _Proxy class is fixed as
well. This certainly is not the best possible solution but for now it
ought to be enough.