fix potential problem with cacheProxy
Until now the cacheProxy class had one problem: There is a potentially
large gap between our system's current time and the server time. When we
receive data from the server its time stamp is aligned based on the
server's time. Now because of the time difference it is possible that we
return data from our cache although there is already more recent data
available on the server.
This change makes use of the timeServer's estimateTime() functionality
to take a guess on the server's current time. With this estimation, we
can now check for the validity of the current cache entry based on the
server's time, not the local one.